Issue #377 - use the VMUSB drivers in place rather than as copies
Import CAEN V830 scaler support directly from vmusb/daqconfig rather
than a hard to maintain copy.
