+++ main/mvlc/Makefile.am
@@ -41,7 +41,8 @@ IMPORTED_MODULES=$(IMPORT_DIR)/CMQDC32RdoHdwr.h \
 	$(IMPORT_DIR)/CMQDC32StackBuilder.h \
 	$(IMPORT_DIR)/CMQDC32StackBuilder.cpp \
 	$(IMPORT_DIR)/MQDC32Registers.h \
-	$(IMPORT_DIR)/C785.cpp $(IMPORT_DIR)/C785.h
+	$(IMPORT_DIR)/C785.cpp $(IMPORT_DIR)/C785.h \
+	$(IMPORT_DIR)/C830.cpp $(IMPORT_DIR)/C830.h
 
 LOCAL_MODULES= CMarker.h CMarker.cpp \
 	C3820.h C3820.cpp \
@@ -53,7 +54,7 @@ LOCAL_MODULES= CMarker.h CMarker.cpp \
 	C3804.h C3804.cpp \
 	CMDPP16QDC.cpp  CMDPP16QDC.h  CMDPP32QDC.cpp  CMDPP32QDC.h \
 	CMDPP32SCP.cpp  CMDPP32SCP.h  CMDPP.h  CMDPPQDC.h \
-	C830.h C830.cpp CNADC2530.h CNADC2530.cpp CV977.h CV977.cpp \
+	CNADC2530.h CNADC2530.cpp CV977.h CV977.cpp \
 	CV1495sc.h CV1495sc.cpp CV1729.h CV1729.cpp \
 	CCAENV1x90Opcodes.h CCAENV1x90Registers.h CV1x90.cpp CV1x90.h \
 	sis3316.h sis3316Registers.h \
